chore (refs T22614): move DpSplitButton to client/js.
